+++ CHANGELOG.md
@@ -24,7 +24,8 @@
     - `ElementLocator.supported_element_attributes_for`
         - Fixed `nil`-error when dealing with unknown attributes.
 - `Session`
-    - Fallback to `Framework` DOM Level 1 handlers when no Browser is available.
+    - Fallback to `Framework` DOM Level 1 handlers when no `Browser` is available.
+    - When `OptionGroups::Scope#dom_depth_limit` is 0 don't use the `Browser`.
 - `Element::Capabilities`
     - `Analyzable`
         - `Differential` -- Forcibly disable `OptionGroups::Audit#cookies_extensively`.